pyproject.toml
[project]
name = "graphmasal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "distro>=1.9.0",
    "graphviz>=0.20.3",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.0.10",
    "langsmith>=0.2.0",
    "neo4j>=6.1.0",
    "networkx>=3.6.1",
    "pydantic>=2.12.5",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.44.0",
]
